+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Send notification to potential research group candidates [#298](https://github.com/rokwire/groups-building-block/issues/298)
 
 ## [1.7.6] - 2022-11-28
 ### Fixed